user catalog tdg tests
